handle UNKNOWN error in isDirectoryAsync graciously
fixes Nexus-Mods/Vortex#2710

not sure why the error code is sometimes UNKNOWN but going by the existing reports it always seems to happen with files, so default to that.
If the file is actually not accessible this will obviously cause another error later on, but maybe we get a better error then.